test: move histogram min/max merge tests into aggregate_test.go
Per review feedback, consolidate the min/max-preservation tests next to
Test_MergeDataPoints in aggregate_test.go instead of a separate
aggregate_minmax_test.go file. No behavior change; the existing file
already imports require and pmetric.
